refaktorointia

+++ tiralabra/src/main/java/tiralabra/util/Const.java
@@ -6,5 +6,5 @@ package tiralabra.util;
  */
 public class Const {
     public static final int pointWidth = 8;
-    public static final double roundPrecision = 100;
+    public static final int roundPrecision = 100;
 }
